update: move todo functions, add selected date functionality
