+++ packages/mgt-components/src/components/mgt-todo/mgt-todo.scss
@@ -51,6 +51,11 @@
     }
   }
 
+  fluent-calendar {
+    align-self: center;
+    width: 300px;
+  }
+
   fluent-text-field {
     width: 800px;
